requirements.txt
Pillow==9.1.1
tk==0.1.0